moving loading to pages
